changes game engine

+++ Source/Timer.h
@@ -1,7 +1,11 @@
+// Timer.h
+// Christopher Ball
+// gets cpu time and frequency for delta time management
+
 #ifndef _TIMER_H_
 #define _TIMER_H_
 
-#include <Windows.h>
+#include "pch.h"
 
 class Timer
 {
@@ -9,8 +13,8 @@ public:
 	Timer();
 	~Timer();
 
-	UINT64	GetTicks();
-	UINT64	GetFrequency();
+	float	GetTicks(); // get cpu time in ticks
+	float	GetFrequency(); // get cpu frequency
 };
 
 #endif _TIMER_H_
